Fixes loading and testing issues in python 2.6
